DONE JUMP APP With Max baudrate is 56000/115200

+++ workspace/MKL46Z4_Boot/drivers/uart0.h
@@ -10,7 +10,7 @@
 //Define value of fields
 //------------- Struct -----------//
 typedef struct{
-	uint16_t baudrate;
+	uint32_t baudrate;
 	uint8_t samplingRatio;
 }UART0_Config_t;
 //================ SUPPORT ================/
